python/python-parser-x509/requirements.txt
pyopenssl
pyasn1
pyasn1-modules
